Don't hardcode the list of supported database backends in the restapi.
Closes gh-449.
